+++ lib/project_types/theme/messages/messages.rb
@@ -94,10 +94,13 @@ module Theme
             Options:
               {{command:--port=PORT}} Local port to serve theme preview from
               {{command:--poll}}      Force polling to detect file changes
-              {{command:--bind=HOST}} Set which network interface the web server listens on
+              {{command:--http-bind=HOST}} Set which network interface the web server listens on
           HELP
           serve: "Viewing theme…",
           open_fail: "Couldn't open the theme",
+          error: {
+            address_binding_error: "Couldn't bind to localhost. To serve your theme, set a different address with {{command:%s theme serve --http-bind=<address>}}",
+          },
         },
         check: {
           help: <<~HELP,